Splitting LoggingValve code out into base and implementation class.
